+++ superfields/src/main/java/org/vaadin/miki/superfields/dates/SuperDatePicker.java
@@ -13,12 +13,13 @@ import org.vaadin.miki.events.text.TextSelectionNotifier;
 import org.vaadin.miki.markers.CanReceiveSelectionEventsFromClient;
 import org.vaadin.miki.markers.CanSelectText;
 import org.vaadin.miki.markers.WithDatePatternMixin;
-import org.vaadin.miki.markers.WithHelper;
+import org.vaadin.miki.markers.WithHelperMixin;
 import org.vaadin.miki.markers.WithIdMixin;
 import org.vaadin.miki.markers.WithLabelMixin;
 import org.vaadin.miki.markers.WithLocaleMixin;
 import org.vaadin.miki.markers.WithPlaceholderMixin;
 import org.vaadin.miki.markers.WithReceivingSelectionEventsFromClientMixin;
+import org.vaadin.miki.markers.WithTitleMixin;
 import org.vaadin.miki.markers.WithValueMixin;
 import org.vaadin.miki.shared.dates.DatePattern;
 import org.vaadin.miki.shared.text.TextSelectionDelegate;
@@ -27,6 +28,7 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.FormatStyle;
 import java.util.Locale;
+import java.util.Objects;
 
 /**
  * An extension of {@link DatePicker} that handles I18N also on the client side.
@@ -39,10 +41,10 @@ import java.util.Locale;
 public class SuperDatePicker extends DatePicker
         implements CanSelectText, CanReceiveSelectionEventsFromClient, WithReceivingSelectionEventsFromClientMixin<SuperDatePicker>,
                    TextSelectionNotifier<SuperDatePicker>, HasSuperDatePickerI18N,
-                   WithLocaleMixin<SuperDatePicker>, WithLabelMixin<SuperDatePicker>,
+                   WithLocaleMixin<SuperDatePicker>, WithLabelMixin<SuperDatePicker>, WithTitleMixin<SuperDatePicker>,
                    WithPlaceholderMixin<SuperDatePicker>, WithDatePatternMixin<SuperDatePicker>,
                    WithValueMixin<AbstractField.ComponentValueChangeEvent<DatePicker, LocalDate>, LocalDate, SuperDatePicker>,
-                   WithIdMixin<SuperDatePicker>, WithHelper<SuperDatePicker> {
+                   WithIdMixin<SuperDatePicker>, WithHelperMixin<SuperDatePicker> {
 
     private final DatePatternDelegate<SuperDatePicker> datePatternDelegate = new DatePatternDelegate<>(this);
 
@@ -192,4 +194,15 @@ public class SuperDatePicker extends DatePicker
             i18n = SuperDatePickerI18nHelper.from(i18n, this.getLocale());
         super.setI18n(i18n);
     }
+
+    @Override
+    public void setTitle(String title) {
+        this.getElement().setProperty("title", Objects.requireNonNullElse(title, ""));
+    }
+
+    @Override
+    public String getTitle() {
+        return Objects.requireNonNullElse(this.getElement().getProperty("title"), "");
+    }
+
 }